app/requirements.txt
opentelemetry-distro==0.58b0
opentelemetry-exporter-otlp==1.37.0
opentelemetry-instrumentation-mysqlclient==0.58b0
opentelemetry-instrumentation-sqlite3==0.58b0
opentelemetry-instrumentation-elasticsearch==0.58b0
opentelemetry-instrumentation-requests==0.58b0
opentelemetry-instrumentation-django==0.58b0
opentelemetry-instrumentation-celery==0.58b0
opentelemetry-instrumentation-httpx==0.58b0
opentelemetry-instrumentation-confluent-kafka==0.58b0
opentelemetry-instrumentation-pymemcache==0.58b0
opentelemetry-instrumentation-logging==0.58b0
opentelemetry-instrumentation-system-metrics==0.58b0
opentelemetry-instrumentation-threading==0.58b0
opentelemetry-instrumentation-wsgi==0.58b0
opentelemetry-instrumentation-asgi==0.58b0
opentelemetry-exporter-prometheus==0.58b0
opentelemetry-semantic-conventions==0.58b0
